refactor: `common-exception` crate should be independent of `meta-types`
The `databend-common-exception` crate, which is a general-purpose
library, should not have a dependency on the business-specific
`databend-common-meta-types` crate.

With this commit, we've relocated the conversion logic for transforming
`Meta**` errors into `ErrorCode`. These conversions are now handled
within the meta-service related crates: `databend-common-meta-types` and
`databend-common-sto-error`.
