+++ src/common/exception/src/exception_into.rs
@@ -18,10 +18,6 @@ use std::fmt::Display;
 use std::fmt::Formatter;
 use std::sync::Arc;
 
-use databend_common_meta_stoerr::MetaStorageError;
-use databend_common_meta_types::MetaAPIError;
-use databend_common_meta_types::MetaError;
-
 use crate::exception::ErrorCodeBacktrace;
 use crate::exception_backtrace::capture;
 use crate::ErrorCode;
@@ -325,24 +321,6 @@ impl From<tonic::Status> for ErrorCode {
             }
             _ => ErrorCode::Unimplemented(status.to_string()),
         }
-    }
-}
-
-impl From<MetaError> for ErrorCode {
-    fn from(e: MetaError) -> Self {
-        ErrorCode::MetaServiceError(e.to_string())
-    }
-}
-
-impl From<MetaAPIError> for ErrorCode {
-    fn from(e: MetaAPIError) -> Self {
-        ErrorCode::MetaServiceError(e.to_string())
-    }
-}
-
-impl From<MetaStorageError> for ErrorCode {
-    fn from(e: MetaStorageError) -> Self {
-        ErrorCode::MetaServiceError(e.to_string())
     }
 }
